+++ core/src/main/java/org/apache/struts/upload/MultipartRequestHandler.java
@@ -20,20 +20,22 @@
  */
 package org.apache.struts.upload;
 
+import java.util.HashMap;
+import java.util.List;
+
 import org.apache.struts.action.ActionMapping;
 import org.apache.struts.action.ActionServlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 
-import java.util.Hashtable;
-
 /**
- * <p> MultipartRequestHandler provides an standard interface for struts to
- * deal with file uploads from forms with enctypes of "multipart/form-data".
- * Providers must provide a no-argument constructor for initialization. </p>
+ * MultipartRequestHandler provides an standard interface for struts to deal
+ * with file uploads from forms with enctypes of "multipart/form-data".
+ * Providers must provide a no-argument constructor for initialization.
  */
 public interface MultipartRequestHandler {
+
     /**
      * <p> This is the ServletRequest attribute that should be set when a
      * multipart request is being read and the maximum length is exceeded. The
@@ -47,31 +49,31 @@ public interface MultipartRequestHandler {
         "org.apache.struts.upload.MaxLengthExceeded";
 
     /**
-     * <p> Convienience method to set a reference to a working ActionServlet
-     * instance. </p>
+     * Convenience method to set a reference to a working ActionServlet
+     * instance.
      */
     public void setServlet(ActionServlet servlet);
 
     /**
-     * <p> Convienience method to set a reference to a working ActionMapping
-     * instance. </p>
+     * Convenience method to set a reference to a working ActionMapping
+     * instance.
      */
     public void setMapping(ActionMapping mapping);
 
     /**
-     * <p> Get the ActionServlet instance </p>
+     * Get the ActionServlet instance.
      */
     public ActionServlet getServlet();
 
     /**
-     * <p> Get the ActionMapping instance for this request </p>
+     * Get the ActionMapping instance for this request.
      */
     public ActionMapping getMapping();
 
     /**
-     * <p> After constructed, this is the first method called on by
-     * ActionServlet. Use this method for all your data-parsing of the
-     * ServletInputStream in the request </p>
+     * After constructed, this is the first method called on by ActionServlet.
+     * Use this method for all your data-parsing of the ServletInputStream in
+     * the request.
      *
      * @throws ServletException thrown if something goes wrong
      */
@@ -79,47 +81,49 @@ public interface MultipartRequestHandler {
         throws ServletException;
 
     /**
-     * <p> This method is called on to retrieve all the text input elements of
-     * the request. </p>
+     * This method is called on to retrieve all the text input elements of the
+     * request.
      *
-     * @return A Hashtable where the keys and values are the names and values
-     *         of the request input parameters
+     * @return A HashMap where the keys and values are the names and values of
+     *         the request input parameters
      */
-    public Hashtable<String, Object> getTextElements();
+    public HashMap<String, String[]> getTextElements();
 
     /**
-     * <p> This method is called on to retrieve all the FormFile input
-     * elements of the request. </p>
+     * This method is called on to retrieve all the FormFile input elements of
+     * the request.
      *
-     * @return A Hashtable where the keys are the input names of the files and
-     *         the values are FormFile objects
+     * @return A HashMap where the keys are the input names of the files and
+     *         the values are FormFile objects.
+     *
      * @see FormFile
      */
-    public Hashtable<String, Object> getFileElements();
+    public HashMap<String, List<FormFile>> getFileElements();
 
     /**
-     * <p> This method returns all elements of a multipart request. </p>
+     * This method returns all elements of a multipart request.
      *
-     * @return A Hashtable where the keys are input names and values are
-     *         either String arrays or FormFiles
+     * @return A HashMap where the keys are input names and values are either
+     *         String arrays or FormFiles.
      */
-    public Hashtable<String, Object> getAllElements();
+    public HashMap<String, Object> getAllElements();
 
     /**
-     * <p> This method is called on when there's some sort of problem and the
-     * form post needs to be rolled back.  Providers should remove any
-     * FormFiles used to hold information by setting them to null and also
-     * physically delete them if the implementation calls for writing directly
-     * to disk. NOTE: Currently implemented but not automatically supported,
+     * This method is called on when there's some sort of problem and the form
+     * post needs to be rolled back. Providers should remove any FormFiles used
+     * to hold information by setting them to null and also physically delete
+     * them if the implementation calls for writing directly to disk.
+     *
+     * <p>NOTE: Currently implemented but not automatically supported,
      * ActionForm implementors must call rollback() manually for rolling back
-     * file uploads. </p>
+     * file uploads.</p>
      */
     public void rollback();
 
     /**
-     * <p> This method is called on when a successful form post has been made.
-     * Some implementations will use this to destroy temporary files or write
-     * to a database or something of that nature. </p>
+     * This method is called on when a successful form post has been made. Some
+     * implementations will use this to destroy temporary files or write to a
+     * database or something of that nature.
      */
     public void finish();
 }